tackled two small bugs (CO2 is now adding 1.5 every time as it should, and formatted the displayed text for saved animals and co2)

+++ app/src/main/java/marrit/marritleenstra_pset6_1/User.java
@@ -84,7 +84,7 @@ public class User implements Serializable {
         return mCO2Avoided;
     }
 
-    public void setCO2Avoided(int CO2Avoided) {
+    public void setCO2Avoided(double CO2Avoided) {
         mCO2Avoided = CO2Avoided;
     }
 }
